Added manual bender control and elevator override
Added manual override for elevator raising and lowering, as well as manual control for the bender, currents just aternates between ampspot and speakerspot, manual control doesn't work?

Added lowering of note when preping amp, this part is not yet tested, as well as the options for auto, also not tested.